requirements.txt
astroid==1.4.8
bleach==1.4.3
cffi==1.8.2
coverage==4.2
cryptography==1.5
Django==1.10.1
django-braces==1.9.0
django-debug-toolbar==1.5
django-markdownx==1.6.3
django-pusherable==0.1.0
django-registration==2.1.2
flake8==3.0.4
gunicorn==19.6.0
html5lib==0.9999999
idna==2.1
isort==4.2.5
lazy-object-proxy==1.2.2
Markdown==2.6.6
markdown2==2.3.1
mccabe==0.5.2
ndg-httpsclient==0.4.2
Pillow==3.3.1
pusher==1.5
pyasn1==0.1.9
pycodestyle==2.0.0
pycparser==2.14
pyflakes==1.2.3
pylint==1.6.4
pyOpenSSL==16.1.0
requests==2.11.1
six==1.10.0
sqlparse==0.2.1
urllib3==1.17
wheel==0.24.0
wrapt==1.10.8
